+++ gtest/coverage.py
@@ -2,36 +2,39 @@
 """
 Test parsing coverage
 
-Usage: gtest (C|coverage) [--skeletons=DIR]
-                          (--list-profiles | <test-pattern> ...)
+Usage: gtest (C|coverage) [--profiles=DIR] [--static]
+                          [--list-profiles]
+                          [<test-pattern> ...]
 
-Arguments (RELPATH: {skeletons}):
-  <test-pattern>        path or glob-pattern to a test skeleton
+Arguments (RELPATH: {profiles}):
+  <test-pattern>        path or glob-pattern to a test skeleton or profile
 
 Options (RELPATH: {grammar-dir}):
-  --skeletons=DIR       skeleton dir [default: :tsdb/skeletons/]
+  --profiles=DIR        profile or skeleton dir [default: :tsdb/skeletons/]
+  -s, --static          don't parse; do static analysis of parsed profiles
   -l, --list-profiles   don't test, just list testable profiles
 
 Examples:
     gtest -G ~/mygram C --list-profiles
     gtest -G ~/mygram C :abc
+    gtest -G ~/mygram C --profiles=:tsdb/gold --static
 """
 
-import re
+from __future__ import print_function
+
 from functools import partial
-from os.path import (join as pjoin, basename, normpath, sep)
-from subprocess import CalledProcessError
+from os.path import join as pjoin
 
 from gtest.util import (
     prepare_working_directory, prepare_compiled_grammar,
     debug, info, warning, error, red, green, yellow,
-    check_exist, make_keypath, dir_is_profile,
-    mkprof, run_art
+    make_keypath, dir_is_profile
 )
 
 from gtest.skeletons import (
-    find_profiles, prepare_profile_keypaths,
-    print_profile_header
+    prepare_profile_keypaths,
+    print_profile_header,
+    test_iterator
 )
 
 from delphin import itsdb
@@ -43,10 +46,9 @@ GENERATE_GOOD = 0.8
 GENERATE_OK = 0.5
 
 def run(args):
-    args['--skeletons'] = make_keypath(args['--skeletons'], args['--grammar-dir'])
-
-    profile_match = partial(dir_is_profile, skeleton=True)
-    prepare_profile_keypaths(args, args['--skeletons'].path, profile_match)
+    args['--profiles'] = make_keypath(args['--profiles'],args['--grammar-dir'])
+    profile_match = partial(dir_is_profile, skeleton=(not args['--static']))
+    prepare_profile_keypaths(args, args['--profiles'].path, profile_match)
 
     if args['--list-profiles']:
         print('\n'.join(map(lambda p: '{}\t{}'.format(p.key, p.path),
@@ -58,66 +60,19 @@ def run(args):
 
 def prepare(args):
     prepare_working_directory(args)
-    with open(pjoin(args['--working-dir'], 'ace.log'), 'w') as ace_log:
-        prepare_compiled_grammar(args, ace_log=ace_log)
+    if not args['--static']:
+        with open(pjoin(args['--working-dir'], 'ace.log'), 'w') as ace_log:
+            prepare_compiled_grammar(args, ace_log=ace_log)
 
 
 def coverage_test(args):
-    for skel in args['<test-pattern>']:
-        name = skel.key
-        logf = pjoin(
-            args['--working-dir'],
-            'run-{}.log'.format(
-                '_'.join(normpath(re.sub(r'^:', '', name)).split(sep))
-            )
-        )
+    for test in test_iterator(args):
+        info('Coverage testing profile: {}'.format(test.name))
         
-        print_profile_header(name, skel.path)
-
-        with open(logf, 'w') as logfile:
-            try:
-                cov = test_coverage(skel, args, logfile)
-                print_coverage_summary(name, cov)
-            except CalledProcessError:
-                print('  There was an error processing the testsuite.')
-                print('  See {}'.format(logf))
+        print_profile_header(test.name, test.destination)
+        cov = parsing_coverage(test.destination)
+        print_coverage_summary(test.name, cov)
             
-
-def test_coverage(skel, args, logfile):
-    info('Coverage testing profile: {}'.format(skel.key))
-
-    cov = {}
-    dest = pjoin(args['--working-dir'], basename(skel.path))
-
-    if not (check_exist(skel.path)):
-        print('  Skeleton was not found: {}'.format(skel.path))
-        return
-
-    mkprof(skel.path, dest, log=logfile)
-    run_art(
-        args['--compiled-grammar'].path,
-        dest,
-        options=args['--art-opts'],
-        ace_preprocessor=args['--preprocessor'],
-        ace_options=args['--ace-opts'],
-        log=logfile
-    )
-
-    cov = parsing_coverage(dest)
-
-    # if args['--generate']:
-    #     g_dest = pjoin(args['--working-dir'], basename(skel.path) + '.g')
-    #     mkprof(skel.path, g_dest, log=logfile)
-    #     run_art(
-    #         args['--compiled-grammar'].path,
-    #         g_dest,
-    #         options=args['--art-opts'] + ['-e', dest],
-    #         ace_preprocessor=args['--preprocessor'],
-    #         ace_options=args['--ace-opts'] + ['-e'],
-    #         log=logfile
-    #     )
-    #     cov = generation_coverage(g_dest, cov)
-    return cov
 
 def parsing_coverage(prof_path):
     # todo: consider i-wf